Modernization Update
- Minimum OS now set to iOS 12
- Modern Launch Screen added (removed black bars at top and bottom)
- Black Screen bug fixed

+++ BreadWallet/BRTxDetailViewController.m
@@ -214,8 +214,8 @@
                         detailLabel.text = NSLocalizedString(@"transaction is post-dated", nil);
                     }
                     else if (peerCount == 0 || relayCount < peerCount) {
-                        detailLabel.text = [NSString stringWithFormat:NSLocalizedString(@"seen by %d of %d peers", nil),
-                                            relayCount, peerCount];
+                        detailLabel.text = [NSString stringWithFormat:NSLocalizedString(@"seen by %lu of %lu peers", nil),
+                                            (unsigned long)relayCount, (unsigned long)peerCount];
                     }
                     else detailLabel.text = NSLocalizedString(@"waiting for confirmation", nil);
